Add pagination to guide pages
Each page in a guide has next page and previous navigation in the footer. This commit also searches the current directory of the rendered page to generate the pagination links using the front matter page_number and page_title to create the HTML.

The markup for previous links is suppressed when the page_number equals 1 and the markup for next link is suppressed when page_number equals the total number of pages in the guide.